+++ src/view/map/LockedCapsule.tsx
